requirements.txt
Django
djangorestframework
drf-yasg==1.21.4

psycopg2-binary==2.9.5

redis==4.4.0
rq==1.11.0

pytest==7.2.0
pytest-django==4.5.2